requirements.txt

# python -m venv venv
# source venv/bin/activate
# pip3 install -r requirements.txt

openslide-python
opencv-python
torch
torchvision
matplotlib
